requirements.txt
gevent  # On M1 Mac, needs to be installed first with `arch -arm64 pip install gevent`
toml
bcrypt
fasteners
GitPython
git+https://github.com/ryanvilbrandt/bottle.git@master
git+https://github.com/ryanvilbrandt/python-markdown2.git@master
bottle_websocket
google-cloud-storage